move vpc/subnet code into separate file
