+++ input/assets_aws/vpc_test.go
@@ -1,0 +1,1 @@
+package assets_aws
